// src/features/workspaces/storage/types/preview.types.ts
export type CrossrefWork = {
    title: string;
    authors: string[];
    editors?: string[];
    doi: string;
    journal: string;
    publicationTitle?: string;
    publicationDate?: string;
    publisher: string;
    place?: string;
    issn: string;
    isbn: string;
    volume: string;
    issue: string;
    section?: string;
    partNumber?: string;
    partTitle?: string;
    pages: string;
    series?: string;
    seriesTitle?: string;
    seriesText?: string;
    year: number | string;
    type: string;
    itemType?: string;
    abstract: string;
    url: string;
    score: number;
    language?: string;
    journalAbbr?: string;
    shortTitle?: string;
    rights?: string;
    license?: string;
    libraryCatalog?: string;
    keywords?: string[];
    pmid?: string;
    pmcid?: string;
    extra?: string;
};

export type PdfMetadata = {
  title?: string;
  author?: string;
  subject?: string;
  creator?: string;
  producer?: string;
  creationDate?: string;
  modDate?: string;
  pageCount?: number;
  keywords?: string;
  doi?: string;
  journal?: string;
  publisher?: string;
  issn?: string;
  isbn?: string;
  volume?: string;
  issue?: string;
  pages?: string;
  publicationDate?: string;
  abstract?: string;
  language?: string;
  copyright?: string;
  year?: number | string;
  authors?: string[];
  editors?: string[];
  type?: string;
  itemType?: string;
  url?: string;
  crossrefEnriched?: boolean;
  extraFields?: Record<string, string>;
  journalAbbr?: string;
  shortTitle?: string;
  rights?: string;
  license?: string;
  publicationTitle?: string;
  place?: string;
  keywordsList?: string[];
};